Added remaining overloads (PATCH, HEAD, OPTIONS, DELETE) Fixed sync extensions to use IRestClient
