+++ src/RestSharp/RestClient.Extensions.cs
@@ -14,82 +14,14 @@
 
 using System.Runtime.CompilerServices;
 using RestSharp.Extensions;
-using RestSharp.Serializers;
 
 namespace RestSharp;
 
 [PublicAPI]
 public static partial class RestClientExtensions {
-    /// <summary>
-    /// Executes a GET-style request asynchronously, authenticating if needed.
-    /// The response content then gets deserialized to T.
-    /// </summary>
-    /// <typeparam name="T">Target deserialization type</typeparam>
-    /// <param name="client"></param>
-    /// <param name="request">Request to be executed</param>
-    /// <param name="cancellationToken">Cancellation token</param>
-    /// <returns>Deserialized response content</returns>
-    public static Task<RestResponse<T>> ExecuteGetAsync<T>(this IRestClient client, RestRequest request, CancellationToken cancellationToken = default)
-        => client.ExecuteAsync<T>(request, Method.Get, cancellationToken);
-
-    /// <summary>
-    /// Executes a GET-style asynchronously, authenticating if needed
-    /// </summary>
-    /// <param name="client"></param>
-    /// <param name="request">Request to be executed</param>
-    /// <param name="cancellationToken">Cancellation token</param>
-    public static Task<RestResponse> ExecuteGetAsync(this IRestClient client, RestRequest request, CancellationToken cancellationToken = default)
-        => client.ExecuteAsync(request, Method.Get, cancellationToken);
-
-    /// <summary>
-    /// Executes a POST-style request asynchronously, authenticating if needed.
-    /// The response content then gets deserialized to T.
-    /// </summary>
-    /// <typeparam name="T">Target deserialization type</typeparam>
-    /// <param name="client"></param>
-    /// <param name="request">Request to be executed</param>
-    /// <param name="cancellationToken">The cancellation token</param>
-    /// <returns>Deserialized response content</returns>
-    public static Task<RestResponse<T>> ExecutePostAsync<T>(
-        this IRestClient  client,
-        RestRequest       request,
-        CancellationToken cancellationToken = default
-    )
-        => client.ExecuteAsync<T>(request, Method.Post, cancellationToken);
-
-    /// <summary>
-    /// Executes a POST-style asynchronously, authenticating if needed
-    /// </summary>
-    /// <param name="client"></param>
-    /// <param name="request">Request to be executed</param>
-    /// <param name="cancellationToken">Cancellation token</param>
-    public static Task<RestResponse> ExecutePostAsync(this IRestClient client, RestRequest request, CancellationToken cancellationToken = default)
-        => client.ExecuteAsync(request, Method.Post, cancellationToken);
-
-    /// <summary>
-    /// Executes a PUT-style request asynchronously, authenticating if needed.
-    /// The response content then gets deserialized to T.
-    /// </summary>
-    /// <typeparam name="T">Target deserialization type</typeparam>
-    /// <param name="client"></param>
-    /// <param name="request">Request to be executed</param>
-    /// <param name="cancellationToken">The cancellation token</param>
-    /// <returns>Deserialized response content</returns>
-    public static Task<RestResponse<T>> ExecutePutAsync<T>(
-        this IRestClient  client,
-        RestRequest       request,
-        CancellationToken cancellationToken = default
-    )
-        => client.ExecuteAsync<T>(request, Method.Put, cancellationToken);
-
-    /// <summary>
-    /// Executes a PUP-style asynchronously, authenticating if needed
-    /// </summary>
-    /// <param name="client"></param>
-    /// <param name="request">Request to be executed</param>
-    /// <param name="cancellationToken">Cancellation token</param>
-    public static Task<RestResponse> ExecutePutAsync(this IRestClient client, RestRequest request, CancellationToken cancellationToken = default)
-        => client.ExecuteAsync(request, Method.Put, cancellationToken);
+    [PublicAPI]
+    public static RestResponse<T> Deserialize<T>(this IRestClient client, RestResponse response)
+        => client.Serializers.Deserialize<T>(response.Request, response, client.Options);
 
     /// <summary>
     /// Executes the request asynchronously, authenticating if needed
@@ -103,12 +35,20 @@ public static partial class RestClientExtensions {
         RestRequest       request,
         CancellationToken cancellationToken = default
     ) {
-        if (request == null)
-            throw new ArgumentNullException(nameof(request));
+        if (request == null) throw new ArgumentNullException(nameof(request));
 
         var response = await client.ExecuteAsync(request, cancellationToken).ConfigureAwait(false);
         return client.Serializers.Deserialize<T>(request, response, client.Options);
     }
+
+    /// <summary>
+    /// Executes the request synchronously, authenticating if needed
+    /// </summary>
+    /// <typeparam name="T">Target deserialization type</typeparam>
+    /// <param name="client"></param>
+    /// <param name="request">Request to be executed</param>
+    public static RestResponse<T> Execute<T>(this IRestClient client, RestRequest request)
+        => AsyncHelpers.RunSync(() => client.ExecuteAsync<T>(request));
 
     /// <summary>
     /// Executes the request asynchronously, authenticating if needed
@@ -128,6 +68,15 @@ public static partial class RestClientExtensions {
         request.Method = httpMethod;
         return client.ExecuteAsync(request, cancellationToken);
     }
+
+    /// <summary>
+    /// Executes the request synchronously, authenticating if needed
+    /// </summary>
+    /// <param name="client"></param>
+    /// <param name="request">Request to be executed</param>
+    /// <param name="httpMethod">Override the request method</param>
+    public static RestResponse Execute(this IRestClient client, RestRequest request, Method httpMethod)
+        => AsyncHelpers.RunSync(() => client.ExecuteAsync(request, httpMethod));
 
     /// <summary>
     /// Executes the request asynchronously, authenticating if needed
@@ -150,140 +99,14 @@ public static partial class RestClientExtensions {
     }
 
     /// <summary>
-    /// Execute the request using GET HTTP method. Exception will be thrown if the request does not succeed.
-    /// The response data is deserialized to the Data property of the returned response object.
+    /// Executes the request synchronously, authenticating if needed
     /// </summary>
-    /// <param name="client">RestClient instance</param>
-    /// <param name="request">The request</param>
-    /// <param name="cancellationToken">Cancellation token</param>
-    /// <typeparam name="T">Expected result type</typeparam>
-    /// <returns></returns>
-    public static async Task<T?> GetAsync<T>(this IRestClient client, RestRequest request, CancellationToken cancellationToken = default) {
-        var response = await client.ExecuteGetAsync<T>(request, cancellationToken).ConfigureAwait(false);
-        return response.ThrowIfError().Data;
-    }
-    
-    public static async Task<RestResponse> GetAsync(this IRestClient client, RestRequest request, CancellationToken cancellationToken = default) {
-        var response = await client.ExecuteGetAsync(request, cancellationToken).ConfigureAwait(false);
-        return response.ThrowIfError();
-    }
-
-    /// <summary>
-    /// Execute the request using POST HTTP method. Exception will be thrown if the request does not succeed.
-    /// The response data is deserialized to the Data property of the returned response object.
-    /// </summary>
-    /// <param name="client">RestClient instance</param>
-    /// <param name="request">The request</param>
-    /// <param name="cancellationToken">Cancellation token</param>
-    /// <typeparam name="T">Expected result type</typeparam>
-    /// <returns></returns>
-    public static async Task<T?> PostAsync<T>(this IRestClient client, RestRequest request, CancellationToken cancellationToken = default) {
-        var response = await client.ExecutePostAsync<T>(request, cancellationToken).ConfigureAwait(false);
-        return response.ThrowIfError().Data;
-    }
-
-    public static RestResponse Post(this IRestClient client, RestRequest request)
-        => AsyncHelpers.RunSync(() => client.PostAsync(request));
-
-    public static async Task<RestResponse> PostAsync(this IRestClient client, RestRequest request, CancellationToken cancellationToken = default) {
-        var response = await client.ExecutePostAsync(request, cancellationToken).ConfigureAwait(false);
-        return response.ThrowIfError();
-    }
-
-    /// <summary>
-    /// Execute the request using PUT HTTP method. Exception will be thrown if the request does not succeed.
-    /// The response data is deserialized to the Data property of the returned response object.
-    /// </summary>
-    /// <param name="client">RestClient instance</param>
-    /// <param name="request">The request</param>
-    /// <param name="cancellationToken">Cancellation token</param>
-    /// <typeparam name="T">Expected result type</typeparam>
-    /// <returns></returns>
-    public static async Task<T?> PutAsync<T>(this IRestClient client, RestRequest request, CancellationToken cancellationToken = default) {
-        var response = await client.ExecuteAsync<T>(request, Method.Put, cancellationToken).ConfigureAwait(false);
-        return response.ThrowIfError().Data;
-    }
-
-    public static async Task<RestResponse> PutAsync(this IRestClient client, RestRequest request, CancellationToken cancellationToken = default) {
-        var response = await client.ExecuteAsync(request, Method.Put, cancellationToken).ConfigureAwait(false);
-        return response.ThrowIfError();
-    }
-
-    /// <summary>
-    /// Execute the request using HEAD HTTP method. Exception will be thrown if the request does not succeed.
-    /// The response data is deserialized to the Data property of the returned response object.
-    /// </summary>
-    /// <param name="client">RestClient instance</param>
-    /// <param name="request">The request</param>
-    /// <param name="cancellationToken">Cancellation token</param>
-    /// <typeparam name="T">Expected result type</typeparam>
-    /// <returns></returns>
-    public static async Task<T?> HeadAsync<T>(this IRestClient client, RestRequest request, CancellationToken cancellationToken = default) {
-        var response = await client.ExecuteAsync<T>(request, Method.Head, cancellationToken).ConfigureAwait(false);
-        return response.ThrowIfError().Data;
-    }
-
-    public static async Task<RestResponse> HeadAsync(this IRestClient client, RestRequest request, CancellationToken cancellationToken = default) {
-        var response = await client.ExecuteAsync(request, Method.Head, cancellationToken).ConfigureAwait(false);
-        return response.ThrowIfError();
-    }
-
-    /// <summary>
-    /// Execute the request using OPTIONS HTTP method. Exception will be thrown if the request does not succeed.
-    /// The response data is deserialized to the Data property of the returned response object.
-    /// </summary>
-    /// <param name="client">RestClient instance</param>
-    /// <param name="request">The request</param>
-    /// <param name="cancellationToken">Cancellation token</param>
-    /// <typeparam name="T">Expected result type</typeparam>
-    /// <returns></returns>
-    public static async Task<T?> OptionsAsync<T>(this IRestClient client, RestRequest request, CancellationToken cancellationToken = default) {
-        var response = await client.ExecuteAsync<T>(request, Method.Options, cancellationToken).ConfigureAwait(false);
-        return response.ThrowIfError().Data;
-    }
-
-    public static async Task<RestResponse> OptionsAsync(this IRestClient client, RestRequest request, CancellationToken cancellationToken = default) {
-        var response = await client.ExecuteAsync(request, Method.Options, cancellationToken).ConfigureAwait(false);
-        return response.ThrowIfError();
-    }
-
-    /// <summary>
-    /// Execute the request using PATCH HTTP method. Exception will be thrown if the request does not succeed.
-    /// The response data is deserialized to the Data property of the returned response object.
-    /// </summary>
-    /// <param name="client">RestClient instance</param>
-    /// <param name="request">The request</param>
-    /// <param name="cancellationToken">Cancellation token</param>
-    /// <typeparam name="T">Expected result type</typeparam>
-    /// <returns></returns>
-    public static async Task<T?> PatchAsync<T>(this IRestClient client, RestRequest request, CancellationToken cancellationToken = default) {
-        var response = await client.ExecuteAsync<T>(request, Method.Patch, cancellationToken).ConfigureAwait(false);
-        return response.ThrowIfError().Data;
-    }
-
-    public static async Task<RestResponse> PatchAsync(this IRestClient client, RestRequest request, CancellationToken cancellationToken = default) {
-        var response = await client.ExecuteAsync(request, Method.Patch, cancellationToken).ConfigureAwait(false);
-        return response.ThrowIfError();
-    }
-
-    /// <summary>
-    /// Execute the request using DELETE HTTP method. Exception will be thrown if the request does not succeed.
-    /// The response data is deserialized to the Data property of the returned response object.
-    /// </summary>
-    /// <param name="client">RestClient instance</param>
-    /// <param name="request">The request</param>
-    /// <param name="cancellationToken">Cancellation token</param>
-    /// <typeparam name="T">Expected result type</typeparam>
-    /// <returns></returns>
-    public static async Task<T?> DeleteAsync<T>(this IRestClient client, RestRequest request, CancellationToken cancellationToken = default) {
-        var response = await client.ExecuteAsync<T>(request, Method.Delete, cancellationToken).ConfigureAwait(false);
-        return response.ThrowIfError().Data;
-    }
-
-    public static async Task<RestResponse> DeleteAsync(this IRestClient client, RestRequest request, CancellationToken cancellationToken = default) {
-        var response = await client.ExecuteAsync(request, Method.Delete, cancellationToken).ConfigureAwait(false);
-        return response.ThrowIfError();
-    }
+    /// <typeparam name="T">Target deserialization type</typeparam>
+    /// <param name="client"></param>
+    /// <param name="request">Request to be executed</param>
+    /// <param name="httpMethod">Override the request method</param>
+    public static RestResponse<T> Execute<T>(this IRestClient client, RestRequest request, Method httpMethod)
+        => AsyncHelpers.RunSync(() => client.ExecuteAsync<T>(request, httpMethod));
 
     /// <summary>
     /// A specialized method to download files.
@@ -301,6 +124,25 @@ public static partial class RestClientExtensions {
 #endif
         return stream == null ? null : await stream.ReadAsBytes(cancellationToken).ConfigureAwait(false);
     }
+
+    /// <summary>
+    /// A specialized method to download files as streams.
+    /// </summary>
+    /// <param name="client"></param>
+    /// <param name="request">Pre-configured request instance.</param>
+    /// <param name="cancellationToken">The cancellation token</param>
+    /// <returns>The downloaded stream.</returns>
+    [PublicAPI]
+    public static Stream? DownloadStream(this IRestClient client, RestRequest request, CancellationToken cancellationToken = default)
+        => AsyncHelpers.RunSync(() => client.DownloadStreamAsync(request, cancellationToken));
+
+    /// <summary>
+    /// A specialized method to download files.
+    /// </summary>
+    /// <param name="client">RestClient instance</param>
+    /// <param name="request">Pre-configured request instance.</param>
+    /// <returns>The downloaded file.</returns>
+    public static byte[]? DownloadData(this IRestClient client, RestRequest request) => AsyncHelpers.RunSync(() => client.DownloadDataAsync(request));
 
     /// <summary>
     /// Reads a stream returned by the specified endpoint, deserializes each line to JSON and returns each object asynchronously.
@@ -342,5 +184,4 @@ public static partial class RestClientExtensions {
             yield return serializer.Deserializer.Deserialize<T>(response)!;
         }
     }
-
 }
