+++ src/RestSharp/RestSharp.csproj
@@ -17,9 +17,6 @@
         <PackageReference Include="Nullable" Version="1.3.1" PrivateAssets="All" />
     </ItemGroup>
     <ItemGroup>
-        <Compile Update="RestClient.Extensions.Config.cs">
-            <DependentUpon>RestClient.Extensions.cs</DependentUpon>
-        </Compile>
         <Compile Update="RestClient.Extensions.Json.cs">
             <DependentUpon>RestClient.Extensions.cs</DependentUpon>
         </Compile>
@@ -34,6 +31,27 @@
         </Compile>
         <Compile Update="Request\PropertyCache.Populator.RequestProperty.cs">
           <DependentUpon>PropertyCache.cs</DependentUpon>
+        </Compile>
+        <Compile Update="RestClient.Extensions.Get.cs">
+          <DependentUpon>RestClient.Extensions.cs</DependentUpon>
+        </Compile>
+        <Compile Update="RestClient.Extensions.Post.cs">
+          <DependentUpon>RestClient.Extensions.cs</DependentUpon>
+        </Compile>
+        <Compile Update="RestClient.Extensions.Put.cs">
+          <DependentUpon>RestClient.Extensions.cs</DependentUpon>
+        </Compile>
+        <Compile Update="RestClient.Extensions.Patch.cs">
+          <DependentUpon>RestClient.Extensions.cs</DependentUpon>
+        </Compile>
+        <Compile Update="RestClient.Extensions.Head.cs">
+          <DependentUpon>RestClient.Extensions.cs</DependentUpon>
+        </Compile>
+        <Compile Update="RestClient.Extensions.Delete.cs">
+          <DependentUpon>RestClient.Extensions.cs</DependentUpon>
+        </Compile>
+        <Compile Update="RestClient.Extensions.Options.cs">
+          <DependentUpon>RestClient.Extensions.cs</DependentUpon>
         </Compile>
     </ItemGroup>
     <ItemGroup>
